Player Status

+++ Assets/Scripts/NavigationUI/RegionStatusScriptable.cs
@@ -6,5 +6,5 @@ public class RegionStatusScriptable : ScriptableObject
     public string worldName;
     public int stageNumber;
     public bool isUnlocked = false, isSelected = false;
-    [Range(0, 3)] public float clearLevel;
+    [Range(0, 3)] public int clearLevel;
 }
